# ui/static/swagger.yaml
openapi: 3.0.0
info:
  title: Garment Recognition API
  description: An API for accessing the Garment Recognition AI model
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Garment Recognition API
tags:
  - name: Garment Recognition Model
    description: Endpoints for accessing the AI model
  - name: Database Operations
    description: Endpoints for accessing the database
paths:
  /search:
    post:
      tags:
        - Garment Recognition Model
      summary: Search for a garment by image
      description: Search for a garment using an image URL
      requestBody:
        description: Search for a garment
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                size:
                  type: integer
              example:
                url: "https://image.hm.com/assets/hm/ea/d7/ead79a8422df6e29abb8e0057c7dbf2d6658bf4c.jpg"
                size: 5
      responses:
        "200":
          description: All the matching garments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Garment"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /keyword_search:
    post:
      tags:
        - Garment Recognition Model
      summary: Search garments by keywords
      description: Search garments by keywords
      requestBody:
        description: Keywords to search for garments
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                keywords:
                  type: array
                  items:
                    type: string
                size:
                  type: integer
              example:
                keywords: ["shirt", "red"]
                size: 5
      responses:
        "200":
          description: List of matching garments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Garment"

  /items/{id}:
    get:
      tags:
        - Database Operations
      summary: Find a garment by ID
      description: Find a garment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The garment information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Garment"
        "404":
          description: Garment not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    delete:
      tags:
        - Database Operations
      summary: Delete a garment by ID
      description: Delete a garment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Garment deleted successfully
        "404":
          description: Garment not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /add_item:
    post:
      tags:
        - Database Operations
      summary: Add a new garment
      description: Add a new garment
      requestBody:
        description: Garment to add
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                url:
                  type: string
                description:
                  type: string
              example:
                name: "Black pants"
                imageUrl: "https://image.hm.com/assets/hm/ea/d7/ead79a8422df6e29abb8e0057c7dbf2d6658bf4c.jpg"
                description: "A pair of black pants"
      responses:
        "201":
          description: Garment added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Garment"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

components:
  schemas:
    Garment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        keywordDescriptions:
          type: array
          items:
            type: string